Refactor: 타입 정의를 room.type.ts, snapshot.type.ts, comment.type.ts, vote.type.ts로 분리
