+++ app/api/rooms/route.ts
@@ -1,23 +1,5 @@
 import { NextRequest, NextResponse } from "next/server";
-
-/**
- * 방 생성 요청 DTO
- */
-export interface CreateRoomRequestDTO {
-  title: string;
-  password: string;
-}
-
-/**
- * 방 생성 응답 DTO
- */
-export interface CreateRoomResponseDTO {
-  roomId: number;
-  uuid: string;
-  title: string;
-  sharedUrl: string;
-  createdAt: string;
-}
+import { CreateRoomRequestDTO, CreateRoomResponseDTO } from "@/types/room.type";
 
 /**
  * 방 생성 요청
